+++ spec/mongoid/criteria_spec.rb
@@ -1408,13 +1408,6 @@ describe Mongoid::Criteria do
             end
           end
 
-          it "does not eager load the last document" do
-            doc = criteria.last
-            expect_query(1) do
-              expect(doc.person).to eq(person_two)
-            end
-          end
-
           it "returns the first document" do
             expect(document).to eq(post_one)
           end
@@ -1423,7 +1416,7 @@ describe Mongoid::Criteria do
         context "when calling last" do
 
           let!(:criteria) do
-            Post.includes(:person)
+            Post.asc(:_id).includes(:person)
           end
 
           let!(:document) do